timer2 baud rate ticks calculation implemented

+++ include/uart.h
@@ -22,7 +22,7 @@
  * 
  * @ingroup uart
  */
-uint16_t uart1_init(uint16_t uart_baudrate);
+uint16_t uart1_init(uint32_t uart_baudrate);
 
 /**
  * Send byte to UART
